requirements.txt
numpy>=1.19.0, <1.20
packaging>=20.4, <20.5
pandas>=1.1.3, <1.2
pluggy>=0.13.1, <0.14
py>=1.9.0, <1.10
PyOpenGL>=3.1.5, <3.2
pyparsing>=2.4.7, <2.5
PyQt5>=5.15.1, <5.16
pyqtgraph>=0.11.0, <0.12
pytest>=6.1.1, <6.2
python-dateutil>=2.8.1, <2.9
pytz>=2020.1, <2020.2
rasterio>=1.2.10, <1.3
scipy>=1.5.2, <1.6
snuggs>=1.4.7, <1.5
toml>=0.10.1, <0.11
xarray>=0.16.1, <0.17
